Get one-line statement blocks working, in line with specified grammar :)

+++ PA3/Tests/PA3Expressions.java
@@ -12,7 +12,7 @@ class PA3Expressions {
 	        Meggy.setPixel( (byte)(2+1-3), (byte)(0+(3+5)) , Meggy.Color.VIOLET );
 		Meggy.setPixel( (byte)(2+(1-3)), (byte)(0-(3+5)) , Meggy.Color.VIOLET );
 	        // Integer subtraction
-	        Meggy.setPixel( (byte)(7-1*6), (byte)(9-5), Meggy.Color.BLUE);
+		Meggy.setPixel( (byte)(7-(byte)1*(byte)6), (byte)(9-5), Meggy.Color.BLUE);
 
 	        // Byte addition
 	        Meggy.setPixel( (byte)((byte)((byte)3+(byte)2)+(byte)1), 
@@ -20,11 +20,11 @@ class PA3Expressions {
 
 	        // Byte subtraction
 	        Meggy.setPixel( (byte)((byte)((byte)3-(byte)2)-(byte)1), 
-	                        (byte)((byte)((byte)0-(byte)7)+(byte)7),
+	                       (byte)((byte)((byte)0-(byte)7)+(byte)7),
 	                        Meggy.Color.YELLOW);
 
                 // Unary Minus
-	        Meggy.setPixel(-5+5, (byte)(10*-5), Meggy.Color.DARK);
+	        Meggy.setPixel(-5+5, (byte)((byte)10*(byte)-5), Meggy.Color.DARK);
                 // watch it! spaces significant
 Meggy.setPixel( (byte)( (byte)(0 - - - 6) * (byte)(-1)), (byte)(10+5), Meggy.Color.ORANGE);
